v2.4

+++ Frontend/Models/ProductModel.cs
@@ -5,7 +5,7 @@
         public int ProductId { get;  set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public decimal? CurrentPrice { get; set; }
+        public double? CurrentPrice { get; set; }
         public int QuantityInStock { get; set; }
     }
 }
